Add information on how to use the shared data
